Extract shell detection to its own package
