+++ commands/shell/shell_test.go
@@ -1,4 +1,4 @@
-package commands
+package shell
 
 import (
 	"os"
@@ -11,7 +11,7 @@ func TestDetectBash(t *testing.T) {
 	originalShell := os.Getenv("SHELL")
 	os.Setenv("SHELL", "/bin/bash")
 	defer os.Setenv("SHELL", originalShell)
-	shell, err := detectShell()
+	shell, err := Detect()
 	assert.Nil(t, err)
 	assert.Equal(t, "bash", shell)
 }
@@ -23,7 +23,7 @@ func TestDetectFish(t *testing.T) {
 	originalFishdir := os.Getenv("__fish_bin_dir")
 	os.Setenv("__fish_bin_dir", "/usr/local/Cellar/fish/2.2.0/bin")
 	defer os.Setenv("__fish_bin_dir", originalFishdir)
-	shell, err := detectShell()
+	shell, err := Detect()
 	assert.Nil(t, err)
 	assert.Equal(t, "fish", shell)
 }
